interpolator and cinematic camera

+++ CosmosRendererLib/CosmosRenderer.h
@@ -102,7 +102,7 @@ public:
 
     const double scale = 0.01;
 
-    float exposure = 0.0003;
+    double exposure = 0.0003;
 
     glm::dvec3 observerCameraPosition;
     double closestSurfaceDistance;
